Wield removed from SMG as intrinsic property
Wieldable component taken off of WT550 and Atreides, they feel pretty strong as a result and could potentially do with a tweak on their performance to prevent them becoming forefront weapons again.

Tested Typewriter without wield component and found it quite enjoyable, but am drawn on implementing it that way.